+++ src/blas/matrix/scal/dense_scal.cpp
@@ -8,7 +8,8 @@ void blas::mscal(const double alpha, matrix::Dense<double> &A) {
   Logger &logger = Logger::get_instance();
   logger.func_in(monolish_func);
 
-  internal::vmul(A.get_nnz(), A.val.data(), alpha, A.val.data(), A.get_device_mem_stat());
+  internal::vmul(A.get_nnz(), A.val.data(), alpha, A.val.data(),
+                 A.get_device_mem_stat());
 
   logger.func_out();
 }
@@ -18,7 +19,8 @@ void blas::mscal(const float alpha, matrix::Dense<float> &A) {
   Logger &logger = Logger::get_instance();
   logger.func_in(monolish_func);
 
-  internal::vmul(A.get_nnz(), A.val.data(), alpha, A.val.data(), A.get_device_mem_stat());
+  internal::vmul(A.get_nnz(), A.val.data(), alpha, A.val.data(),
+                 A.get_device_mem_stat());
 
   logger.func_out();
 }